test: add integration test assemblies
Integration tests for the Admin API are run sequentially, using a web
app fixture that is reset to an empty state after every test.

Integration tests for the Public API are run in parallel, using a web
app fixture that is to be seeded with test data before any tests are
run. This is possible because all the Public API endpoints are to be
GET only.

Closes: #43