test: add integration test assemblies
Added two integration test assemblies, one for each API assembly.

It should be noted that:

1. Integration tests for the Admin API are run sequentially using a web
app fixture that is initially empty and is reset to its initial state
after every test.

2. Integration tests for the Public API (which will only ever have GET
endpoints) are run in parallel using a web app fixture that is seeded
before the first test is run.

Closes: #45