feat(shared): add problem details responses
Every request enters the app pipeline and returns as either a successful
response or a list of errors.

If the pipeline returns a successful response, the endpoint maps the
response to a successful HTTP response containing the response object.

If the pipeline returns a list of errors, the endpoint maps the errors
to an unsuccessful HTTP response containing a `ProblemDetails` object
converted from the first error.

Closes: #49