feat(shared): add api key security
Implemented API key authentication and role-based authorization for the
Admin API and the Public API.

Two API keys are defined: a demo API key and a secret API key.

Requests to the Admin API must use the secret API key as an 'X-Api-Key'
request header. Requests to the Public API must use either the secret
API key or the demo API key.

Closes: #43
Closes: #44
Closes: #45