feat(admin-api): add contest created handling
Added a `ContestCreatedEvent` to the domain.

Updated the `Entity` base class to include a domain events collection.

Updated the EF Core persistence so that any domain events are
intercepted and published as part of the same transaction.

When a new contest is created using the Admin API, every participating
country updates itself with a reference to the created contest.

Closes: #18