feat(admin-api): add contest creation handling
When a contest is created, all the countries involved in the contest
are automatically updated with a contest role referencing the contest.

This commit introduces domain events to the system.

Domain events are published by the `BaseWriteRepository` abstract base
class before all changes including those caused by event handlers are
saved to the database in a single transaction.

Closes: #19