chore: migrate to .net 10
Updated the entire solution to use the newly-released .NET 10.

As part of this commit, the 'global.json' file was updated to specify
that the Microsoft Testing Platform should be used with the
`dotnet test` command.

Also as part of this commit, the following Microsoft packages were all
updated to version 10.0.0 -

- Microsoft.AspNetCore.Mvc.Testing
- Microsoft.AspNetCore.OpenApi
- Microsoft.EntityFrameworkCore
- Microsoft.EntityFrameworkCore.Design
- Microsoft.EntityFrameworkCore.SqlServer

All tests still pass.